# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Laravel Documentation",
        "description": "Blog API",
        "contact": {
            "name": "Amikel Maxi",
            "email": "maxloversist@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "summary": "Create a new user",
                "operationId": "ca7c0cb6772de9bffbb8c896980cebc8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User and the user token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validations input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "summary": "Log in the user",
                "operationId": "c373bc4cffa4a56fc64e3979b8ee7b96",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User object and Token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "Token",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User credentials incorrect"
                    },
                    "422": {
                        "description": "Input check validations message"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "summary": "Logged out the user (revoke the token)",
                "operationId": "5ff2dc8b48bff52908d3821a0ecc9b3e",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    },
                    "401": {
                        "description": "Invalid token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/add/comments": {
            "post": {
                "summary": "Create a new Comment",
                "operationId": "23619da747570efb2cf87f0a43e25ea3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Comment"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comment created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Post id is required or Comment is required or Post id must be an integer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/add/likes": {
            "post": {
                "summary": "Create a new like",
                "operationId": "5bc8891d4e31b87fb278eb74a640dc5d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Like"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Liked successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Like"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Post id is required or Post Id must be an integer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Like"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "You can like a post only once!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Like"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/remove/likes/{id}": {
            "put": {
                "summary": "Remove a like to a post",
                "operationId": "d5b97e77d68e0435861a768cc8a49852",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Liked successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Like"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Comment": {
                "required": [
                    "comment"
                ],
                "properties": {
                    "id": {
                        "description": "Comment id",
                        "type": "integer"
                    },
                    "comment": {
                        "description": "Comment content",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Like": {
                "required": [
                    "like"
                ],
                "properties": {
                    "id": {
                        "description": "The like id",
                        "type": "integer"
                    },
                    "like": {
                        "description": "Like (true or false)",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
...
            "User": {
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "id": {
                        "description": "The User ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The User name",
                        "type": "string"
                    },
                    "email": {
                        "description": "The User email",
                        "type": "string"
                    },
                    "password": {
                        "description": "The User password",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}